ofz#7006 tweak to avoid signed int overflow

+++ src/lib/MSPUBCollector.cpp
@@ -459,13 +459,13 @@ Coordinate getFudgedCoordinates(Coordinate coord, const std::vector<Line> &lines
   }
   else
   {
-    if (unsigned(fudged.m_xe - fudged.m_xs) > leftFudge)
+    if (int64_t(fudged.m_xe) - fudged.m_xs > leftFudge)
       fudged.m_xs += leftFudge;
-    if (unsigned(fudged.m_xe - fudged.m_xs) > rightFudge)
+    if (int64_t(fudged.m_xe) - fudged.m_xs > rightFudge)
       fudged.m_xe -= rightFudge;
-    if (unsigned(fudged.m_ye - fudged.m_ys) > topFudge)
+    if (int64_t(fudged.m_ye) - fudged.m_ys > topFudge)
       fudged.m_ys += topFudge;
-    if (unsigned(fudged.m_ye - fudged.m_ys) > bottomFudge)
+    if (int64_t(fudged.m_ye) - fudged.m_ys > bottomFudge)
       fudged.m_ye -= bottomFudge;
   }
   return fudged;
